more vec integer cleanup
