+++ hacspec-lib/src/lib.rs
@@ -31,9 +31,6 @@ mod seq;
 mod traits;
 mod transmute;
 mod util;
-mod vec_integers;
-mod vec_integers_public;
-mod vec_integers_secret;
 mod vec_util;
 
 pub use crate::prelude::*;
