+++ hacspec-lib/src/math_util/vec_integers.rs
