+++ hacspec-lib/tests/test_public_seq.rs
@@ -13,18 +13,3 @@ fn test_byte_sequences() {
         assert_eq!(((msg_u32[i] & 0xFF000000) >> 24) as u8, U8::declassify(chunk[0]));
     }
 }
-
-#[test]
-fn test_seq_poly() {
-    let x = PublicSeq::<u128>::from_native_slice(&[5, 2, 7, 8, 9]);
-    let y = PublicSeq::<u128>::from_native_slice(&[2, 1, 0, 2, 4]);
-    let z = x.clone() + y.clone();
-    let z = z.clone() - x.clone();
-    assert_eq!(z[0..y.len()], y[0..y.len()]);
-    assert_eq!(z.iter().skip(y.len()).fold(0, |acc, x| acc+x), 0);
-    let _z = x.clone() * y.clone();
-    // let (zq, zr) = z.clone() / x.clone();
-    // assert_eq!(zr.iter().fold(0, |acc, x| acc+x), 0);
-    // assert_eq!(zq[0..y.len()], y[0..y.len()]);
-    // assert_eq!(zq.iter().skip(y.len()).fold(0, |acc, x| acc+x), 0);
-}
